Move content to newly created http-app file.
